peer: Update Connected Propererty of P2P Technology
Currently the property Connected is not used for P2P Technology thus it
always remains False. This patch aims to keep it updated in order to
make easier for the UI to show if P2P technology has an active P2P
connection or not.

This implementation has a limitation in cases of complex systems where
multiple P2P interfaces are present and can be connected simultaneously,
it is because the disconnection of one of those interfaces will cause
that the property becomes False even if there is another interface still
connected. But let say that such a complex systems are not very common.