network: Remove noisy DBGs and merge one
These DBGs are noisy in the log and don't provide any real
value.

Little point in having two DBGs. We seldom run out of memory. And
if so, we have for sure a more interesting problems. So move
it to the end of the function which makes the log useful again.