service: Allow removing service in idle state
Currently only if the service state is CONNMAN_SERVICE_STATE_FAILURE,
the service is allowed to be removed. This commit allows removing the
service when service state is CONNMAN_SERVICE_STATE_IDLE and
CONNMAN_SERVICE_STATE_DISCONNECT.

This is needed if a connection attempt fails and the service State is
in CONNMAN_SERVICE_STATE_IDLE or CONNMAN_SERVICE_STATE_DISCONNECT,
Remove() method can be used to reset the service to clear any wrong
passphrase given.