service: Set method to AUTO when address is got through IPv4LL
When DHCP fails and ConnMan obtains an address through IPv4LL, the method
stored/notified for that service must be "auto" instead of "dhcp" as has
been done so far. This patch aims to fix that misleading.

However, after a disconnection from a service with "auto" as IPv4 method,
ConnMan should use DHCP by first when user tries to reconnects or
auto-connect starts on that service. Same reasoning applies after a
power-cycle, i.e. when ConnMan starts up and reads "auto" as method from
the settings file, it should also use DHCP by first and only if it fails
then IPv4LL should be used.

The IPv4.Configuration is also set to "auto" in order to allow user to ask
ConnMan to try to get an address through DHCP at any time by manually
setting this property to "dhcp". On the other hand, "auto" cannot be set
by users from D-Bus API, as it has been so far.