wifi: Do not allow scan/autoscan on p2p connected interface
In eed27a6e1d1 was introduced the concept of p2p_device to identify the
dedicated P2P-Group interface created by wpa_supplicant when a P2P
connection is established. This concept is used to deny Scan and Autoscan
(Introduced in b584447dde1) on P2P connected interfaces.

However, if wpa_supplicant is configured to not create such dedicated
P2P-Group interface and instead use directly the same one, then ConnMan
needs a way to continue denying Scan/Autoscan on that P2P connected
interface and that is the scope if this patch. With this modification,
the p2p_device flag is also used to temporally mark the current interface
connected with P2P.