wifi: Make sure to reply scan request when connected with P2P
As analog situation of e04c74b1eba, in this case is also required that
plugin replies error to the scan request otherwise ConnMan will never
emit the D-Bus reply and counterpart application will go in D-Bus time-out.
For instance, this will always happen in a system with only one interface
with wpa_supplicant configured to not create dedicated P2P-Group interface
for P2P connections.