wifi: Set current network in case of reconnection
Use associated notification from gsupplicant to ensure ConnMan is
prepared when wpa_s triggers a connection with the last AP we were
connected to that previously deauthenticated us. If it is not done, WiFi
plugin could receive a notification it does not expect and state
machines will not move, resulting in a L2 connection and no IP address.

This should rarely happen because wpa_s should always notify there is a
new BSS through a "BSSAdded" signal before notifying we got connected to
it. If wpa_s does so, everything will be in its place when wpa_s
notifies we got connected.