pacrunner: Do not configure our domains
There is a confusion between our domains, and the proxy domains.

The domain configured for the service have no link to the proxy
domain.

PacRunner says

	array{string} Domains [optional]

		Domain names and IP range for which this proxy
		configuration shall be valid.

So for example a proxy servers may be valid to access website under
the domain example.com to be able to reply to request for url:
http://www.example.com/site/test.html host: www.example.com but not
for http://ipv4.connman.net/online/status.html ipv4.connman.net.

Just remove this Domain parameter since we don't know for which domain
the proxy server is relevant, assume it's for any domain.