session: Keep track of addr in fw_snat & session
When there is more than one session in fw_snat's list of sessions,
fw_snat failed to be re-created when update-session-state is triggered
with new IP address. This is because index alone is not sufficient to
decide if fw_snat needs to be re-created. The solution here is to keep
a track of IP addr and use it to avoid false lookup of fw_snat.