gsupplicant: Updates to handling of security change
Changing of security type of one BSS is resulting in removing of
GSupplicantNetwork and other BSSs objects in the group.  The
corresponding entries for other BSS in the same group are not removed
in other hash tables like bss_mapping and interface->bss_mapping.
This commit only deletes the network when there is a single entry in
network->bss_table during change in security for one bss.  This commit
also avoids accessing any deleted bss object when reply is got from
async dbus call, by deleting any pending dbus calls associated with
the bss object.