wifi: Fix behaviour when Background scanning is disabled
Conditions when ConnMan will perform active instead of passive scanning:
1. WiFi Driver support active scanning, it means "MaxScanSSID" is
   greater than 1.
2. User has successfully connected to at least one WiFi network in the
   past so the service information is present in
   {STOREDIR}/connman/wifi_* but none of those networks are currently in
   range. They will be called "known-services" from now.

Current behaviour when BackgroundScanning is enabled:
- D-Bus scan request or at start-up of system -> Launch an
  active/passive scan (Based on above conditions)
- 3 seconds -> Launch passive scan
- 9 seconds -> Launch passive scan
- ...
- 243 seconds -> Launch passive scan
- 300 seconds -> Launch passive scan
- Continue each 300 seconds unless user performs a new D-Bus scan call
  which will cause the procedure to restart from 3 seconds.

This patch does not change the ConnMan's behaviour when
BackgroundScanning is enabled. However, when it is disabled and
conditions for active scanning are true, ConnMan could fall in the
following situation:
- D-Bus scan request or at start-up of system -> Launch an active scan
- Stop scanning

One single scan is done and it is an active one. It means that ConnMan
will perform an active scan looking for those known-services in their
specific frequencies, this kind of scanning is generally faster than a
passive one which allows to speed up auto-connection procedure.
However, we have seen that in case of none of these knwon-services is in
range some drivers reply only with all the networks in the specified
frequencies and some other does not reply any network. This could result
in an always empty services list.

Therefore, this patch looks for avoid such situation by adding a passive
scan in such cases. The new behaviour is also based on the conditions
mentioned at the beginning:
1. Active scanning:
- D-Bus scan request or at start-up of system -> Launch an active scan
- 3 seconds -> Launch passive scan
- Stop scanning
- New D-Bus scan request from user -> Launch an active scan
- 3 seconds -> Launch passive scan
- Stop scanning

2. Passive scanning:
- D-Bus scan request or at start-up of system -> Launch one passive scan
- Stop scanning
- New D-Bus scan request from user -> Launch one passive scan
- Stop scanning

Take into account that all the behaviours described above are followed
when wifi has not get connected. Once it is connected, ConnMan stops its
background scanning procedure and if BackgroundScanning is enabled, it
asks wpa_supplicant to perform the background scanning while connected
only for wifi roaming purposes. See "wpa_supplicant configuration"
section in README file for further details.