ofono: Removing contexts requires a careful list walk
The function remove_cm_context() modifies the context_list so the
list->next pointer is no longer valid on its return.

With this patch, we _carefully_ iterate over the context_list, assuming
that it may have been modified on each iteration.

Without this patch, only the first context in the list is successfully
removed; after that the behaviour is somewhat unpredictable!  What I was
seeing was that services coupled to the second context were never
destroyed resulting in available services without a proper backing
device.