connection: Set and unset active flag when changing default gateway
The active flag for gateway should be set also when it is set or unset
as default gateway as it is done with VPN gateways. The active flag is
used in finding the active gateway (find_active_gateway(), choosing the
default gateway (choose_default_gateway()), updating the default gateway
(__connman_connection_update_gateway()) and to delete routes
(disable_gateway()).

The active flag is set/unset when adding or removing the gateway but not
touched when updating the interface using set_default_gateway() and
unset_default_gateway(). This adds setting of the flag in both of the
uses of the set/unset default gateway and unset_default_gateway(). This
adds setting of the flag in both of the uses of the set/unset default
gateway.