openconnect: Rewrite plugin to support more auth methods and protocols
Co-authored-by: Santtu Lakkala <santtu.lakkala@jolla.com>

The purpose of this rewrite is to extend OpenConnect plugin to support
more authentication methods and support settings for them. Also VPN
agent use in the plugin is improved. In order to support self signed
certificate and PKCS pass phrase requests, and to improve detection
of errors stderr is monitored and reacted accordingly. Furthermore, all
the three protocols are supported via provider settings.

Following new option strings were added:
 NAME					TYPE (internal)	VIA
 OpenConnect.AllowSelfSignedCert        string (bool)   Provider
 OpenConnect.AuthType			string		Provider
 OpenConnect.DisableIPv6                string (bool)   Provider
 OpenConnect.NoDTLS                     string (bool)   Provider
 OpenConnect.NoHTTPKeepalive            string (bool)   Provider
 OpenConnect.PKCSClientCert		string (path)   Provider
 OpenConnect.PKCSPassword	        string          Agent
 OpenConnect.Protocol                   string          Provider
 OpenConnect.Usergroup                  string          Provider
 OpenConnect.UserPrivateKey		string (path)	Provider
 VPN.MTU				string (int)	Provider

The supported types for authentication types ("OpenConnect.AuthType")
are:
 - Cookie:					"cookie"
 - Use cookie after username and password auth: "cookie_with_userpass"
 - Username and password:			"userpass"
 - Certificate using public key:		"publickey"
 - Use PKCS#1/#8/#12 file:			"pkcs"

Without defining the authentication type, OpenConnect defaults to "cookie".
The value of the "OpenConnect.AuthType" is saved to provider config and
resulting provider settings file.

In case the "cookie_with_userpass" is used as authentication method
first the username and password are used to get the OpenConnect cookie
(only authentication is performed, OpenConnect is run with --cookieonly)
and then the resulting cookie is used in the next connect. Only with
this authentication method stdout is monitored for cookie string.

With "pkcs" the password is queried via agent for the PKCS file. This
is implemented as separate from the public key authentication as the
separation of the actions required for the authentication are more clear
clear this way. PKCS decryption errors are sent as priority data to
stderr, which must be read in order to detect it. It is imperative to
close the io channel as well since OpenConnect will write a new PKCS
pass phrase request to stderr as well, without newline, which will block
until timeout. It is better to close the channel in case of decryption
error to avoid connman-vpnd freezing.

Username, Password and PKCSPassword are to be retrieved via VPN agent.
These are saved as hidden provider values and are not stored to disk.

ClientCert, UserPrivateKey and PKCSClienCert are all paths to the
corresponding files on local filesystem.

Authentication failures are sent to VPN agent to indicate that saved
credentials should be cleared. For this the changes to auth and
connection error tracking in vpn-provider.c, and the changes to
vpn-agent.c are utilized.

In order to be compatible with the previous version of the plugin the
default authentication, "cookie" is left as is from VPN agent part
except the VPN server after authentication is set as optional. With
other authentication methods server SHA1 fingerprint and VPN server
after authentication are to be set into provider settings.

Option "AllowSelfSignedCert" can be set as string "true" or "false".
The value defaults "false" if value is omitted or invalid. Self signed
server certificates have to be explicitly allowed. If the value is false
the --servercert reported by OpenConnect via stderr is not saved to
provider settings and connection is terminated. Followed by this
OpenConnect requests a reply to indicate that the certificate needs to
be approved, or denied, and depending on the option, answer is written
to stdin ("yes") if self signed certs are allowed and otherwise the
connection is terminated. In case the value is false the task parameter
"--servercert" is also omitted when running the task. In case the value
is true, the self signed certificate is saved, if returned by the
OpenConnect client and the server certificate SHA1 fingerprint is stored
to OpenConnect.ServerCert in provider settings.

The following were added to OpenConnect startup parameters if defined in
provider settings as "true".

PROVIDER SETTING                PARAMETER
OpenConnect.DisableIPv6		--disable-ipv6
OpenConnect.NoDTLS		--no-dtls
OpenConnect.NoHTTPKeepalive	--no-http-keepalive
VPN.MTU				--base-mtu

Additionally, OpenConnect is set to use syslog for all connection related
output with --syslog. This is because with syslog enabled only the
cookie is delivered via stdout.

All the three different protocols supported by OpenConnect can be
selected with provider setting OpenConnect.Protocol.

VALUE			DEFINITION
anyconnect		Cisco AnyConnect (default)
nc			Juniper Network Connect / Pulse Secure
gp			Palo Alto Networks GlobalProtect