l2tp: Save 'exclusive' config value in lns default context
The config value 'exclusive' must be saved in lns default context. This
is because if written to lac context parsing will fail (versions 1.3.8
and 1.3.14 were tested):
xl2tpd[]: parse_config: line 12: 'exclusive' not valid in this context

This went undetected unless the option was set. By default, 'exclusive'
is not written to xl2tpd config. The xl2tpd documentation states that
"Most of those described in the LNS section may be used in a LAC
context, where it make common sense (essentially l2tp procotols tuning
flags and authentication / ppp related ones)."

Added a enum option OPT_L2LNS for defining other options to be used in
lns context only.