vpn-provider: Change provider in failure state to idle when saving
If the VPN provider is in failure state and it is saved the old state
must be cleared and changed to idle. This represents better the VPN
provider state in such scenario, the failure can be caused by connection
error or login error, both of which must be reset if the settings are
saved (most likely user changes settings to get the VPN to work).