provider: Handle -EALREADY in __connman_provider_connect()
The error -EALREADY returned by vpn.c:provider_connect() in case there
is already a pending request for connecting the VPN is to be handled
separately from -EINPROGRESS. With -EALREADY the state has not changed
and should not be indicated. From service.c point of view the returned
status should still be -EINPROGRESS as the actual status of the VPN has
not changed.