iwd: Add support for tethering
Toggle the Mode of the device between 'Station' and
'AccessPoint'. Since ConnMan doesn't handle the tethering per device,
we just pick the first device which supports both modes and try to set
the mode. In case we disable tethering try to set all devices back to
'Station'. For the AccessPoint mode we only enable it on the first
device.