ipconfig: Do not clear ipaddress for fixed and manual configurations
During disconnect disconnected __connman_ipconfig_address_remove()
will be called. This removes the IP configuration from the interface
and also clears the ipconfig->address. For dynamic IP configuration
this is the right thing to do. But for manual and fixed we lose
the current configuration. At this point the settings file is still
written correctly out (the previous addr/mask/gateway). But when
user decided to change any configuration like autoconnect while
the service is in disconnect state, ipconfig will write out an
'empty' settings file.

  connmanctl> config ethernet_e0d55ee31e14_cable --ipv4 manual
    192.168.154.174 255.255.255.0 192.168.154.1
  connmanctl> disconnect ethernet_e0d55ee31e14_cable
  Disconnected ethernet_e0d55ee31e14_cable
  connmanctl> config ethernet_e0d55ee31e14_cable --autoconnect no

will result in

  [ethernet_e0d55ee31e14_cable]
  Name=Wired
  AutoConnect=false
  Modified=2020-01-09T08:58:11Z
  IPv4.method=manual
  IPv6.method=auto
  IPv6.privacy=enabled