wifi: Do no disconnect when disconnect code is unspecified reason
Cisco wireless controllers seem to send the disconnect code 1
sometimes. It could be a very dirty way to get the client moving to a
different AP.

Anyway, the '1' case was added as precaution and doesn't have any real
bug report behind it. So let's drop it as we had two independent
reports that ConnMan is not doing the right thing.