provider: Add support for managing SplitRouting
Support SplitRouting via provider, add a function for plugins to set the
value: connman_provider_set_split_routing(). In to ensure the proper
handling of split routing value changes utilize the separated service
moving __connman_service_move_service() function to reorder the services
when VPN is connected.  When a VPN is to be set as split routed it is
used as a target service, and when VPN is not split routed it should be
regarded as the default service.

If the VPN is not connected only the split routing value is changed for
the service. If there is an error in moving the connected service use
__connman_service_split_routing_changed() to signal vpnd about the
current status of split routing.

To ensure that setting file has the SplitRouting value if changed in VPN
settings by vpnd the vpn_service must be saved. This is to be done
explicitly in provider.c since in service.c the saving is done
explicitly as well after changing the SplitRouting value.

To allow checking the split routing status also in VPN plugin add
connman_provider_is_split_routing() to get the split routing value
from the actual service. By default returns false if the provider or
the service is omitted.

Add getter for provider family: connman_provider_get_family(). Return
AF_UNSPEC with NULL provider.