vpn-provider: Drop route management from vpnd
This change removes the '-r' option to toggle whether to manage routes
or not in connman-vpnd. Remove all functionality from vpn-provider.c
that was used for managing routes. Remove all functions managing the
routes. All routes should be managed by connmand only.

The change was originally introduced in a7851eb04e67 ("vpnd: Add -r
option which enables route handling in vpnd") for the most part. The
initial idea was the VPN daemon runs stand alone and handles the
routing. Since the introduction of the splittet routing, the stand
alone mode was only partially working.

This breaks existing (are there any?) who only run the VPN daemon
without ConnMan. If so, you need to implement the routing table
updates on your end.