wifi: Fix wireless interface not being added to tether bridge sometimes
For Broadcom BCM4356 chipset with brcmfmac driver, the IFF_LOWER_UP flag
did not always change state in wifi_newlink() but remained set when
changing from STA mode to AP mode. This resulted in handle_tethering()
not being called to add the wireless interface to the tether bridge.

To resolve the issue, always call handle_tethering() as long as the
IFF_LOWER_UP flag is set instead of only when IFF_LOWER_UP changes
from unset to set. The handle_tethering() function already has checks
in place to avoid adding the wireless interface to the tether bridge
more than once.