service: Restart online check when default service changes
In a multi-connected networks environment, it is necessary to start
the online check on a new default service. For example, when Ethernet
cable is unplugged and ConnMan connects to a WiFi network the default
service switches from the Ethernet service to the WiFi services. In
this case ConnMan should start the online check for WiFi.