service: Add online check interval config options
Global config options, which allow to set the range of intervals
between two online check requests.
When an online check request fails, another one is triggered after a
longer interval. The intervals follow the power of two series of numbers
between OnlineCheckInitialInterval and OnlineCheckMaxInterval.
Default range is [1, 12], corresponding to the following intervals, in
seconds: 1, 4, 9, 16, 25, 36, 49, 64, 81, 100, 121 and 144.