service: Support hot-plug of technologies by updating ipconfig index
It is important to update the index of the ipconfig belonging to the
service that is to be created from network after re-inserting a
technology that supports hot-plug (e.g., an USB ethernet adapter). The
service does exist in such case but the interface index is not updated
which results in DHCP failing to start. This is because the old index
would be used.

With this change index update is always done regardless of the existing
path for the service.